% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLMMD.R
\name{BLMMD}
\alias{BLMMD}
\title{Genotype x Environment models using linear or gaussian kernel}
\usage{
BLMMD(Y, K, XF=NULL, method=c("GK", "G-BLUP"), h=NULL, nIter = 1000, burnIn = 200, thin = 5, ...)
}
\arguments{
\item{XF}{\code{matrix} Design matrix (\eqn{n \times p}) for fixed effects}

\item{thin}{\code{integer} Thinin interval.}

\item{Y}{\code{data.frame} Phenotypic data with three columns. The first column is a \code{factor} for assigned environments,
the second column is a \code{factor} for assigned individuals and the third column contains the trait of interest.}

\item{nIter}{\code{integer} Number of iterations.}

\item{burnIn}{\code{integer} Number of iterations to be discarded as burn-in.}

\item{\dots}{additional arguments to be passed.}
}
\description{
Genotype x Environment models using linear or gaussian kernel
}
\details{
The goal is to fit genomic prediction models including GxE interaction. These models can be adjusted through two different kernels.



export
}
