% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getK.R
\name{getK}
\alias{getK}
\title{Kernel matrix for GxE models}
\usage{
getK(Y, X, kernel = c("GK", "GB"), K = NULL,
            h = 1, model = c("SM", "MM", "MDs", "MDe", "Cov"))
}
\arguments{
\item{Y}{\code{data.frame} Phenotypic data with three columns. The first column is a factor for environments,
the second column is a factor for genotype identification and the third column contains the trait of interest}

\item{X}{Marker matrix with individuals in rows and marker in columns}

\item{kernel}{Kernel to be used. Methods currently implemented are the gaussian \code{GK} and the linear \code{G-BLUP} kernel}

\item{K}{\code{matrix} Single kernel matrix in case it needs to use a different kernel from those supported}

\item{h}{\code{numeric} Bandwidth parameter to create the Gaussian Kernel (GK) matrix. The default for \code{h} is 1.
Estimation of h can be made using h.fun function}

\item{model}{Specifies the genotype by environment model to be fitted. It currently supported the models  \code{SM}, \code{MM}, \code{MDs} and \code{MDe}. See Details}
}
\value{
It returns a list, which has the suitable kernels for respective model. The possible kernel created are:
\itemize{
\item G: kernel for  main genotypic effect
\item GE:



}
}
\description{
Create kernel matrix for some GxE models R
}
\details{
The goal is to create kernels used to fit GxE interaction models. These models can be adjusted using different kernels.
\code{GB} creates a linear kernel resulted from the cross-product of centered and standarized marker genotypes divide by the number of markers \eqn{p}:
    \deqn{G = \frac{XX^T}{p}}
Other kernel option currently supported is the Gaussian Kernel \code{Gk}, resulted from exponential of a genetic distance matrix based on markers scaled by its fifth percentile multiplied by the bandwidth parameter \eqn{h}.
Thus:
 \deqn{ K (x_i, x_{i'}) = exp(-h d_{ii'}^2)}
However other kernels can be provided through \code{K}. In this case, arguments \code{X}, \code{kernel} and \code{h} are ignored.

The currently supported models for GxE kernels are:
\itemize{
\item \code{SM}: is the single-environment main genotypic effect model - The SM model fits the data for single environment and only one kernel is produced. 
\item \code{MM}: is the multi-environment main genotypic effect model - Multi-environment model considering the main random genetic effects across environments and one kernel is produced, with dimension \eqn{n \times n}, related to main effect across environments. 
\item \code{MDs}: is the multi-environment single variance genotype x environment deviation model - This model is an extension of \code{MM} by adding the random interaction effect of the environments. Thus, two kernels are created, where the first one is related to main effect across environment,
and the second kernel is associated with genotype by enviroment effect.
\item \code{MDe}: is the multi-environment, environment-specific variance genotype x environment deviation model - This model separates the genetic effects of markers into the main marker effects and the specific marker effects (for each environment). Thus, one kernel for across environments effect is created and
\eqn{j} kernels are created, one for each environment. 
\item \code{Cov}: is the multi-environment, variance-covariance environment-specific model - Only one kernel is created with all unique genotypes available.
}
}
\examples{
# create kernel matrix for model MDs using wheat dataset
library(BGLR)

data(wheat)
X <- scale(wheat.X, scale = TRUE, center = TRUE)
rownames(X) <- 1:599
pheno_geno <- data.frame(env = gl(n = 4, k = 599), 
               GID = gl(n=599, k=1, length = 599*4),
               value = as.vector(wheat.Y))
 K <- getK(Y = pheno_geno, X = X, kernel = "GB", model = "MDs")              



}
