% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h.fun.R
\name{h.fun}
\alias{h.fun}
\title{Selection of bandwidth parameter \eqn{h} for kernel regression}
\usage{
h.fun(Y, D)
}
\arguments{
\item{Y}{\code{data.frame} The first column is a factor for environments,
the second column is a factor for genotype identification and the third column contains the trait of interest. Missing values
should be assigned as NA.}

\item{D}{Genetic distance between individuals.}
}
\value{
it returns the value for \eqn{h}
}
\description{
Estimation of bandwidth parameter \eqn{h} of the Gaussian kernel by Bayesian method
}
\details{
The reproducing kernel (RK) function has two components: a genetic distance between individuals based on markers and the bandwidth parameter 
\eqn{h} that controls the rate of decay of the covariance between genotypes. Here, this parameter is estimate from data. 
The approach is based on the bayesian method for selecting the bandwidth parameter \eqn{h} through its marginal distribution. For more details see Perez-Elizalde et al. (2015).
This function uses all dataset available, however missing data are ignored to estimate the \eqn{h}. If necessary to compute one \eqn{h} for each environment a proper subset should be used.
}
\examples{
# get h for one environment
library(BGLR)

data(wheat)
X <- scale(wheat.X, scale = TRUE, center = TRUE)
D <- (as.matrix(dist(X))) ^ 2
rownames(X) <- 1:599
pheno_geno <- data.frame(env = gl(n = 1, k = 599), 
               GID = gl(n=599, k=1),
               value = as.vector(wheat.Y[,1]))
h <- h.fun(Y = pheno_geno, D = D)

}
\references{
Pérez-Elizalde, S.,Cuevas, J.; Pérez-Rodríguez, P.; Crossa, J. (2015) Selection of The Bandwidth Parameter in a Bayesian Kernel Regression Model for Genomic-Enabled Prediction.
J Agr Biol Envir S, 20-4:512–532
}
