% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainGE.R
\name{mainGE}
\alias{mainGE}
\title{Genotype x Environment models using linear or gaussian kernel}
\usage{
mainGE <- function(Y, X, XF=NULL, W=NULL, method=c("GK", "G-BLUP"), h=NULL, model = c("SM", "MM", "MDs", "MDe", "Cov"), nIter = 150, burnIn = 50, thin = 5, ...)
}
\arguments{
\item{Y}{\code{data.frame} Phenotypic data with three columns. The first column is a \code{factor} for assigned environments,
the second column is a \code{factor} for assigned individuals and the third column contains the trait of interest.}

\item{X}{\code{matrix} Marker matrix with individuals in rows and marker in columns}

\item{XF}{\code{matrix} Design matrix (\eqn{n \times p}) for fixed effects}

\item{W}{\code{matrix} Environmental covariance matrix. If \code{W} is provided, it can be used along with \code{MM} and \code{MDs} models. See details}

\item{method}{Kernel to be used. Methods implemented are the gaussian kernel \code{Gk-EB} and the linear kernel \code{G-BLUP}}

\item{h}{\code{numeric} Bandwidth parameter to create the gaussian kernel matrix. For \code{method} \code{Gk-EB}, if \code{h} is not provided,
then it is computed following a empirical bayesian selection method. See details}

\item{model}{Specifies the genotype by environment model to be fitted. \code{SM} is the single-environment main genotypic effect model, 
\code{MM} is the multi-environment main genotypic effect model, \code{MDs} is the multi-environment single variance genotype × environment deviation model,
\code{MDe} is the multi-environment, environment-specific variance genotype × environment deviation model}

\item{nIter}{\code{integer} Number of iterations.}

\item{burnIn}{\code{integer} Number of iterations to be discarded as burn-in.}

\item{thin}{\code{integer} Thinin interval.}

\item{\dots}{additional arguments to be passed.}
}
\value{

}
\description{
Genotype x Environment models using linear or gaussian kernel
}
\examples{


export

}
\seealso{
\code{\link[MTM]{MTM}}, \code{\link[BGLR]{BGLR}} and \code{\link{function}}
}
