% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.BGGE}
\alias{plot.BGGE}
\title{Comparative plot}
\usage{
plot(BGGE_Object, ...)
}
\arguments{
\item{x}{\code{BGGE object}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Simple plot of the predicted values versus observed values
}
