% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.BGGE}
\alias{print.BGGE}
\title{Print BGGE information object}
\usage{
\method{print}{BGGE}(x, ...)
}
\arguments{
\item{x}{BGGE object}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Displays the most relevant model fit information.
}
\description{
Print BGGE information object
}
