% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLMMD.R
\name{BGGE}
\alias{BGGE}
\title{Genotype x Environment models using regression kernel}
\usage{
BGGE(y, K, XF = NULL, ne, ite = 1000, burn = 200, thin = 3, verbose = FALSE, tol = 1e-10)
}
\arguments{
\item{y}{Vector of data. Should be numeric and NAs are allowed.}

\item{K}{list A two-level list Specify the regression kernels (co-variance matrix). The first element is \code{Kernel},
where is included the regression kernels. The second element is the \code{Type}, specifying if the matrix is either \code{D} Dense or
\code{BD} Block Diagonal. A number of regression kernels or random effects to be fitted are specify in this list.}

\item{XF}{matrix Design matrix (\eqn{n \times p}) for fixed effects}

\item{ne}{vector Number of genotypes by environment.}

\item{ite}{numeric Number of iterations.}

\item{burn}{numeric Number of iterations to be discarded as burn-in.}

\item{thin}{numeric Thinin interval.}

\item{verbose}{Should iteration history be printed on console? if TRUE or 1 then it is printed,
otherwise if another number $n$ is choosen the history is printed every $n$ times. Default is FALSE.}

\item{tol}{a numeric tolerance level. Eigenvalues lower than \code{tol} are discarded. Default is 1e-10.}
}
\value{
A list with estimated posterior means of variance components for each term in the linear model and the genetic value predicted. Also along with 
 the chains are released
}
\description{
BGGE function fits Bayesian regression for continuous observations through regression kernels
}
\details{
The goal is to fit genomic prediction models for continuous outcomes through Gibbs sampler. BGGE uses a proposal
for dimension reduction through orthogonal transformation of observed data (y) as well as differential
shrinkage as a result of the prior variance assigned to regression parameters. Further details on this approach can be found in Cuevas et al. (2014).
The basic genetic model is
\deqn{y = g + e}
where \eqn{y} is the response, \eqn{g} is the unknown random effect and \eqn{e} is the residual effect.
You can specify a number of random effects \eqn{g}, as many as desired, through a list with regression kernels related to each random effect in the
argument \code{K}.
The structure of \code{K} is a two level list, where the first element on the second level is the Kernel and the second element is a definition on
type of matrix. There is two definitions, either matrices are \code{D} (dense) or \code{BD} (Block Diagonal). As we make the spectral decomposition 
on the kernels, for block diagonal matrices, we take advantage of its structure and make decomposition on the submatrices instead of one matrices.
For example, the regression kernels should be an structure like K = list(list(Kernel = G, Type = "D"), list(Kernel = G, Type = "BD")). 
The definition of one matrix as a block diagonal must be followed by the number of subjects in each submatrix in the block diagonal,
present in the \code{ne}, which allows sub matrices to be drawn. Some genotype by environment models has the block diagonal matrix type or similar.
The genotype x environment deviation matrix in MDs model (Sousa et al., 2017) has the structure of block diagonal. 
Also, the matrices for environment-specific variance in MDe models (Sousa et al., 2017) if summed, can form an structure of block diagonal, 
where is possible to extract sub matrices for each environment. In case of all kernel be of dense type, \code{ne} is ignored.
}
\examples{
# single environment model fit
library(BGLR)

data(wheat)
X <- scale(wheat.X, scale = TRUE, center = TRUE)
K <- list(G = list(Kernel = tcrossprod(X)/ncol(X), Type = "D"))
y <- as.vector(wheat.Y[,1])

fit <- BGGE(y=y, K=K, ne = 599)

}
\references{
Cuevas, J., Perez-Elizalde, S., Soberanis, V., Perez-Rodriguez, P., Gianola, D., & Crossa, J. (2014).
Bayesian genomic-enabled prediction as an inverse problem. G3: Genes, Genomes, Genetics, 4(10), 1991-2001.

Sousa, M. B., Cuevas, J., Oliveira, E. G. C., Perez-Rodriguez, P., Jarquin, D., Fritsche-Neto, R., Burgueno, J. 
& Crossa, J. (2017). Genomic-enabled prediction in maize using kernel models with genotype x environment interaction.
G3: Genes, Genomes, Genetics, 7(6), 1995-2014.
}
\seealso{
\code{\link[BGLR]{BGLR}}
}
