% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLMMD.R
\name{BLMMD}
\alias{BLMMD}
\title{Genotype x Environment models using linear or gaussian kernel}
\usage{
BLMMD(y, K, XF = NULL, ite = 1000, burn = 200, thin = 3, verbose = FALSE, tol = 1e-10)
}
\arguments{
\item{y}{vector of phenotypic information. Should be numeric type.}

\item{K}{list Specify the regression kernels (co-variance matrix) to be fitted. A number of matrices with regression kernels to be fitted should be
included in this list.}

\item{XF}{matrix Design matrix (\eqn{n \times p}) for fixed effects}

\item{ite}{numeric Number of iterations.}

\item{burn}{numeric Number of iterations to be discarded as burn-in.}

\item{thin}{numeric Thinin interval.}

\item{verbose}{Should iteration history be printed on console? if TRUE or 1 then it is printed,
otherwise if another number $n$ is choosen the history is printed every $n$ times. Default is FALSE.}

\item{tol}{a numeric tolerance level. Eigenvalues lower than \code{tol} are discarded. Default is 1e-10.}
}
\value{
A list with estimated posterior means of variance components for each term in the linear model and the genetic value predicted.
}
\description{
BLMMD function fits Bayesian regression for continuous variables through linear or Gaussian kernel.
}
\details{
The goal is to fit genomic prediction models for continuous outcomes through Gibbs sampler. BLMMD uses a proposal
for dimension reduction through orthogonal transformation of observed data (y) as well as differential
shrinkage as a result of the prior variance assigned to regression parameters. Further details on this approach can be found in Cuevas et al. (2014).
The basic genetic model is
\deqn{y = g + e}
where \eqn{y} is the response, \eqn{g} is the unknown random effect and \eqn{e} is the residual effect.
You can specify a number of random effects \eqn{g}, as many as desired, through a list with regression kernels related to each random effect.
}
\examples{
# single environment model fit
library(BGLR)

data(wheat)
X <- scale(wheat.X, scale = TRUE, center = TRUE)
K <- list(G = tcrossprod(X)/ncol(X))
y <- as.vector(wheat.Y[,1])

fit <- BLMMD(y=y, K=K)

}
\references{
Cuevas, J., Perez-Elizalde, S., Soberanis, V., Perez-Rodriguez, P., Gianola, D., & Crossa, J. (2014).
Bayesian genomic-enabled prediction as an inverse problem. G3: Genes, Genomes, Genetics, 4(10), 1991-2001.
}
\seealso{
\code{\link[BGLR]{BGLR}}
}
