% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLMMD.R
\name{BLMMD}
\alias{BLMMD}
\title{Genotype x Environment models using linear or gaussian kernel}
\usage{
BLMMD(y, K, XF = NULL, ite = 1000, burn = 200, thin = 3, verbose = FALSE, me = 1e-10)
}
\arguments{
\item{y}{vector of phenotypic data}

\item{K}{Kernels with effects to be fitted}

\item{XF}{\code{matrix} Design matrix (\eqn{n \times p}) for fixed effects}

\item{ite}{\code{integer} Number of iterations.}

\item{burn}{\code{integer} Number of iterations to be discarded as burn-in.}

\item{thin}{\code{integer} Thinin interval.}

\item{verbose}{Should iteration history be printed on screen? if TRUE or 1 then it is printed,
otherwise if another number $n$ is choosen the history is printed every $n$ times. Default is FALSE.}

\item{me}{\code{numeric} Trheshold for selecting eigenvalues. Eigenvalues lower than \code{me} are discarded. Default is 1e-10.}
}
\description{
Genotype x Environment models using linear or gaussian kernel
}
\details{
The goal is to fit genomic prediction models including GxE interaction. These models can be adjusted through two different kernels.
}
\examples{
# single environment model fit
library(BGLR)

data(wheat)
X <- scale(wheat.X, scale = TRUE, center = TRUE)
K <- list(G = list(K = tcrossprod(X)/ncol(X)))
y <- as.vector(wheat.Y[,1])

fit <- BLMMD(y=y, K=K)

}
\seealso{
\code{\link[BGLR]{BGLR}}
}
