% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getK.R
\name{getK}
\alias{getK}
\title{Kernel matrix for GxE models}
\usage{
getK(Y, X, kernel = c("GK", "GB"), K = NULL, h = 1,
             model = c("SM", "MM", "MDs", "MDe"), quantil = 0.5)
}
\arguments{
\item{Y}{\code{data.frame} Phenotypic data with three columns. The first column is a factor for environments,
the second column is a factor for indentifying genotypes and the third column contains the trait of interest}

\item{X}{Marker matrix with individuals in rows and markers in columns. Missing markers are not allowed.}

\item{kernel}{Kernel to be created internally. Methods currently implemented are the Gaussian \code{GK} and the linear \code{GBLUP} kernel}

\item{K}{\code{matrix} Single kernel matrix in case it needs to use a different kernel from \code{GK} or \code{GBLUP}}

\item{h}{\code{numeric} Bandwidth parameter to create the Gaussian Kernel (GK) matrix. The default for \code{h} is 1.
Estimation of h can be made using an bayesian approach in Perez-Elizalde et al (2015)}

\item{model}{Specifies the genotype \eqn{\times} environment model to be fitted. It currently supported the models  \code{SM}, \code{MM}, \code{MDs} and \code{MDe}. See Details}

\item{quantil}{Specifies the quantile to create the gaussian kernel}
}
\value{
It returns a list, with suitable kernels for each model. It creates the kernels: \code{G} which is for main genotypic effect, included in all models;
\code{GE} for genotype by environment effect created in the MDs model and for the \code{MDe} for each environment, one kernel is created.
}
\description{
Create kernel matrix for genomic GxE models
}
\details{
The aim is to create kernels to fit GxE interaction models. These models can be fitted using different kernels.
\code{GB} creates a linear kernel resulted from the cross-product of centered and standarized marker genotypes divide by the number of markers \eqn{p}:
    \deqn{G = \frac{XX^T}{p}}
Other kernel option currently supported by the BLMMD is the Gaussian Kernel \code{Gk}, resulted from the exponential of a genetic distance matrix based on markers scaled by its fifth percentile multiplied by the bandwidth parameter \eqn{h}.
Thus:
 \deqn{ K (x_i, x_{i'}) = exp(\frac{-h d_{ii'}^2}{q_{0.05}(d)})}
However other kernels can be provided through \code{K}. In this case, arguments \code{X}, \code{kernel} and \code{h} are ignored.

The currently supported models for GxE kernels are:
\itemize{
\item \code{SM}: is the single-environment main genotypic effect model - The SM model fits the data for single environment and only one kernel is produced. 
\item \code{MM}: is the multi-environment main genotypic effect model - Multi-environment model considering the main random genetic effects across environments and one kernel is produced, with order \eqn{n \times n}, related to main effect across environments. 
\item \code{MDs}: is the multi-environment single variance genotype x environment deviation model - This model is an extension of \code{MM} by adding the random interaction effect of environments with genotype information. Thus, two kernels are created, one related to main effect across environment,
and the second kernel is associated with genotype by enviroment effect.
\item \code{MDe}: is the multi-environment, environment-specific variance genotype x environment deviation model - This model separates the genetic effects into the main genetic effects and the specific genetic effects (for each environment). Thus, one kernel for across environments effect is created and
\eqn{j} kernels are created, one for each environment.
}
}
\examples{
# create kernel matrix for model MDs using wheat dataset
library(BGLR)

data(wheat)
X <- scale(wheat.X, scale = TRUE, center = TRUE)
rownames(X) <- 1:599
pheno_geno <- data.frame(env = gl(n = 4, k = 599), 
               GID = gl(n=599, k=1, length = 599*4),
               value = as.vector(wheat.Y))
 K <- getK(Y = pheno_geno, X = X, kernel = "GB", model = "MDs")              



}
\references{
Jarquín, D., J. Crossa, X. Lacaze, P. Du Cheyron, J. Daucourt, J. Lorgeou, F. Piraux, L. Guerreiro, P. Pérez, M. Calus, J. Burgueño,
and G. de los Campos. 2014. A reaction norm model for genomic selection using high-dimensional genomic and 
environmental data. Theor. Appl. Genet. 127(3): 595–607.

López-Cruz, M., J. Crossa, D. Bonnett, S. Dreisigacker, J. Poland, J.-L. Jannink, R.P. Singh, E. Autrique,
and G. de los Campos. 2015. Increased prediction accuracy in wheat breeding trials using a marker × environment
interaction genomic selection model. G3: Genes, Genomes, Genetics. 5(4): 569–82.
}
