% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h.fun.R
\name{h.fun}
\alias{h.fun}
\title{Genotype x Environment models using linear or gaussian kernel}
\usage{
mainGE <- function(Y, X, XF=NULL, W=NULL, method=c("GK", "G-BLUP"), h=NULL, model = c("SM", "MM", "MDs", "MDe", "Cov"), nIter = 150, burnIn = 50, thin = 5, ...)
}
\arguments{
\item{Y}{\code{data.frame} Phenotypic data with three columns. The first column is a \code{factor} for assigned environments,
the second column is a \code{factor} for assigned individuals and the third column contains the trait of interest.}

\item{X}{\code{matrix} Marker matrix with individuals in rows and marker in columns}
}
\description{
Genotype x Environment models using linear or gaussian kernel
}
\details{
The goal is to estimate the bandwith parameter from data. The approach used is a bayesian method for selecting the bandwidth parameter \eqn{h}
through the marginal distribution of \eqn{h}. For more details see Perez-Elizalde et al. (2015).
export
}
