% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainGE2.R
\name{mainGE}
\alias{mainGE}
\title{Genotype x Environment models using linear or gaussian kernel}
\usage{
mainGE(Y, X, XF=NULL, W=NULL, method=c("GK", "GB"),
              h=1, model = c("SM", "MM", "MDs", "MDe", "Cov"),
              nIter = 1000, burnIn = 300, thin = 5, verbose = FALSE, ...)
}
\arguments{
\item{Y}{\code{data.frame} Phenotypic data with three columns. The first column is a \code{factor} for assigned environments,
the second column is a \code{factor} for assigned individuals and the third column contains the trait of interest.}

\item{X}{\code{matrix} Marker matrix with individuals in rows and marker in columns}

\item{XF}{\code{matrix} Design matrix (\eqn{n \times p}) for fixed effects}

\item{W}{\code{matrix} Environmental covariance matrix of dimension (\eqn{n \times n}). If \code{W} is provided, it can be used along with \code{MM} and \code{MDs} models. See details}

\item{method}{Kernel to be used. Methods implemented are the gaussian kernel \code{Gk-EB} and the linear kernel \code{G-BLUP}}

\item{h}{\code{numeric} Bandwidth parameter to create the gaussian kernel matrix. For \code{method} \code{Gk-EB}, if \code{h} is not provided,
then it is computed following a empirical bayesian selection method. See details}

\item{model}{Specifies the genotype by environment model to be fitted. \code{SM} is the single-environment main genotypic effect model,
\code{MM} is the multi-environment main genotypic effect model, \code{MDs} is the multi-environment single variance genotype × environment deviation model,
\code{MDe} is the multi-environment, environment-specific variance genotype × environment deviation model}

\item{nIter}{\code{integer} Number of iterations.}

\item{burnIn}{\code{integer} Number of iterations to be discarded as burn-in.}

\item{thin}{\code{integer} Thinin interval.}

\item{\dots}{additional arguments to be passed.}
}
\value{
return variance components and
}
\description{
Genotype x Environment models using linear or gaussian kernel
}
\details{
The goal is to fit genomic prediction models including GxE interaction. These models can be adjusted through two different kernels.
\code{G-BLUP} creates a linear kernel resulted from the cross-product of centered and standarized marker genotypes divide by the number of markers \eqn{p}:
    \eqn{G = \frac{XX^T}{p}}
If \code{Gk} is choosen, a gaussian kernel is created, resulted from exponential of a genetic distance matrix based on markers scaled by its fifth percentile multiplied by the bandwidth parameter \eqn{h}
which has an objective of controlling the rate of decay for correlation between individuals. Thus:
 \eqn{ K (x_i, x_{i'}) = exp(-h d_{ii'}^2)}
However if the bandwidth parameter is not provided, it need to be estimated from data. The approach currently working is a bayesian method for selecting the bandwidth parameter \eqn{h}
through the marginal distribution of \eqn{h}. For more details see Perez-Elizalde et al. (2015).
mainGE uses the packages BGLR and MTM to fit the current models:
\begin{itemize}
\item \code{SM}: is the single-environment main genotypic effect model - The SM model fits the data for each single environment separately.
\item \code{MM}: is the multi-environment main genotypic effect model - Multi-environment model considering the main random genetic effects across environments.
\item \code{MDs}: is the multi-environment single variance genotype × environment deviation model - This model is an extension of \code{MM} by adding the random interaction effect of the environments
with the genetic information of genotypes.
\item \code{MDe}: is the multi-environment, environment-specific variance genotype × environment deviation model - This model separates the genetic effects of markers into the main marker effects (across environments) and the specific marker effects (for each environment).
\item \code{Cov}: is the multi-environment, variance-covariance environment-specific model -
}
\examples{
# create kernel matrix for model MDs using wheat dataset
library(BGLR)

data(wheat)
X <- scale(wheat.X, scale = T, center = T)
rownames(X) <- 1:599
pheno_geno <- data.frame(env = gl(n = 4, k = 599), 
               GID = gl(n=599, k=1, length = 599*4),
               value = as.vector(wheat.Y))

fit <- mainGE(Y = pheno_geno, X = X, method = GB, model = "MDs")
 

export
}
\seealso{
\code{\link[MTM]{MTM}}, \code{\link{getK}} and \code{\link{BLMMD}}
}
